Move OTEL stuff to `firezone-telemetry`
