+++ rust/telemetry/Cargo.toml
@@ -6,6 +6,8 @@ license = { workspace = true }
 
 [dependencies]
 anyhow = { workspace = true }
+ip-packet = { workspace = true }
+opentelemetry = { workspace = true }
 parking_lot = { workspace = true }
 reqwest = { workspace = true }
 sentry = { workspace = true, features = ["contexts", "backtrace", "debug-images", "panic", "reqwest", "rustls", "tracing"] }
